# apps/api/src/swagger/api-schema.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/game/bosses": {
      "get": {
        "operationId": "GameController_getBosses",
        "parameters": [
          {
            "name": "lang",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "pl",
                "en",
                "de"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicBossResponse"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Game"
        ]
      }
    },
...
    "/game/admin/bosses": {
      "get": {
        "operationId": "GameController_getAdminBosses",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Game"
        ]
      },
      "post": {
        "operationId": "GameController_createBoss",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBossBody"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Game"
        ]
      }
    },
    "/game/admin/stats": {
      "get": {
        "operationId": "GameController_getAdminStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameAdminStatsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Game"
        ]
      }
    },
    "/game/admin/import": {
      "post": {
        "operationId": "GameController_importGameContent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportGameContentBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportGameContentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Game"
        ]
      }
    },
    "/game/admin/bosses/{uuid}": {
      "patch": {
        "operationId": "GameController_updateBoss",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBossBody"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Game"
        ]
      },
      "delete": {
        "operationId": "GameController_deleteBoss",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Game"
        ]
      }
    },
    "/game/admin/questions": {
      "get": {
        "operationId": "GameController_getAdminQuestions",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Game"
        ]
      },
      "post": {
        "operationId": "GameController_createQuestion",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionBody"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Game"
        ]
      }
    },
    "/game/admin/questions/{uuid}": {
      "patch": {
        "operationId": "GameController_updateQuestion",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestionBody"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Game"
        ]
      },
      "delete": {
        "operationId": "GameController_deleteQuestion",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Game"
        ]
      }
    }
  },
  "info": {
    "title": "Portfolio website API",
    "description": "API documentation for the portfolio website",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "TranslatedTextDto": {
        "type": "object",
        "properties": {
          "pl": {
            "type": "string"
          },
          "en": {
            "type": "string"
          },
          "de": {
            "type": "string"
          }
        },
        "required": [
          "pl",
          "en",
          "de"
        ]
      },
...
      "PublicBossResponse": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "maxRounds": {
            "type": "number"
          },
          "questionCount": {
            "type": "number"
          }
        },
        "required": [
          "uuid",
          "slug",
          "name",
          "description",
          "maxRounds",
          "questionCount"
        ]
      },
      "GameAdminStatsResponse": {
        "type": "object",
        "properties": {
          "activePlayers": {
            "type": "number",
            "description": "Number of players currently connected to active game rooms",
            "example": 4,
            "minimum": 0
          }
        },
        "required": [
          "activePlayers"
        ]
      },
      "CreateBossBody": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/TranslatedTextDto"
          },
          "description": {
            "$ref": "#/components/schemas/TranslatedTextDto"
          },
          "maxRounds": {
            "type": "number",
            "minimum": 5,
            "maximum": 30
          },
          "published": {
            "type": "boolean"
          }
        },
        "required": [
          "slug",
          "name",
          "description",
          "maxRounds"
        ]
      },
      "GameAnswerBody": {
        "type": "object",
        "properties": {
          "content": {
            "$ref": "#/components/schemas/TranslatedTextDto"
          },
          "isCorrect": {
            "type": "boolean"
          },
          "orderIndex": {
            "type": "number"
          },
          "matchKey": {
            "type": "string"
          }
        },
        "required": [
          "content",
          "isCorrect"
        ]
      },
      "ImportQuestionBody": {
        "type": "object",
        "properties": {
          "content": {
            "$ref": "#/components/schemas/TranslatedTextDto"
          },
          "explanation": {
            "$ref": "#/components/schemas/TranslatedTextDto"
          },
          "type": {
            "type": "string",
            "enum": [
              "SINGLE_CHOICE",
              "TRUE_FALSE",
              "ORDER",
              "MATCHING"
            ]
          },
          "category": {
            "type": "string"
          },
          "difficulty": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "sourceUrl": {
            "type": "string"
          },
          "verifiedAt": {
            "type": "string"
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameAnswerBody"
            }
          },
          "published": {
            "type": "boolean"
          }
        },
        "required": [
          "content",
          "explanation",
          "type",
          "category",
          "difficulty",
          "sourceUrl",
          "verifiedAt",
          "answers"
        ]
      },
      "ImportBossBody": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/TranslatedTextDto"
          },
          "description": {
            "$ref": "#/components/schemas/TranslatedTextDto"
          },
          "maxRounds": {
            "type": "number",
            "minimum": 5,
            "maximum": 30
          },
          "published": {
            "type": "boolean"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportQuestionBody"
            }
          }
        },
        "required": [
          "slug",
          "name",
          "description",
          "maxRounds",
          "questions"
        ]
      },
      "ImportGameContentBody": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "enum": [
              1
            ],
            "example": 1
          },
          "bosses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportBossBody"
            }
          }
        },
        "required": [
          "version",
          "bosses"
        ]
      },
      "ImportedBossResponse": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "uuid",
          "slug"
        ]
      },
      "ImportGameContentResponse": {
        "type": "object",
        "properties": {
          "bossesImported": {
            "type": "number"
          },
          "questionsImported": {
            "type": "number"
          },
          "bosses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportedBossResponse"
            }
          }
        },
        "required": [
          "bossesImported",
          "questionsImported",
          "bosses"
        ]
      },
      "UpdateBossBody": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/TranslatedTextDto"
          },
          "description": {
            "$ref": "#/components/schemas/TranslatedTextDto"
          },
          "maxRounds": {
            "type": "number",
            "minimum": 5,
            "maximum": 30
          },
          "published": {
            "type": "boolean"
          }
        }
      },
      "CreateQuestionBody": {
        "type": "object",
        "properties": {
          "bossUuid": {
            "type": "string"
          },
          "content": {
            "$ref": "#/components/schemas/TranslatedTextDto"
          },
          "explanation": {
            "$ref": "#/components/schemas/TranslatedTextDto"
          },
          "type": {
            "type": "string",
            "enum": [
              "SINGLE_CHOICE",
              "TRUE_FALSE",
              "ORDER",
              "MATCHING"
            ]
          },
          "category": {
            "type": "string"
          },
          "difficulty": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "sourceUrl": {
            "type": "string"
          },
          "verifiedAt": {
            "type": "string"
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameAnswerBody"
            }
          },
          "published": {
            "type": "boolean"
          }
        },
        "required": [
          "bossUuid",
          "content",
          "explanation",
          "type",
          "category",
          "difficulty",
          "sourceUrl",
          "verifiedAt",
          "answers"
        ]
      },
      "UpdateQuestionBody": {
        "type": "object",
        "properties": {
          "bossUuid": {
            "type": "string"
          },
          "content": {
            "$ref": "#/components/schemas/TranslatedTextDto"
          },
          "explanation": {
            "$ref": "#/components/schemas/TranslatedTextDto"
          },
          "type": {
            "type": "string",
            "enum": [
              "SINGLE_CHOICE",
              "TRUE_FALSE",
              "ORDER",
              "MATCHING"
            ]
          },
          "category": {
            "type": "string"
          },
          "difficulty": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "sourceUrl": {
            "type": "string"
          },
          "verifiedAt": {
            "type": "string"
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameAnswerBody"
            }
          },
          "published": {
            "type": "boolean"
          }
        }
      }
    }
  }
}